fix(box-measure): QA fixes — threading, nav, compile errors
ARCore threading (critical):
- ArCoreBoxMeasureView: Session.update() was called on the UI thread inside
  setOnTouchListener, concurrent with GL thread's onDrawFrame. Rewrote to use
  ConcurrentLinkedQueue for tap events; hit-testing now runs exclusively inside
  onDrawFrame on the GL thread (ArRenderer inner class).
- Session.close() now dispatched via GLSurfaceView.queueEvent (closeOnGlThread)
  to prevent main-thread race with onDrawFrame.

Navigation (important):
- ShiftNavGraph: onDimensionsVerified was calling popBackStack(), but
  BoxMeasureScreen already calls onBack() after the callback — double-pop
  removed. Verified dims now stored in previousBackStackEntry.savedStateHandle
  for future POP Track-A payload wiring.
- app/build.gradle.kts: added :feature:box-measure dependency (BoxMeasureScreen
  was unresolved from ShiftNavGraph).

Expo module:
- ArMeasurementModule (Android): concurrent measureBox() calls now reject the
  previous pending promise before accepting a new one, preventing indefinite hangs.

Compile fixes:
- ArCoreBoxMeasureView: FloatArray element subtraction returns Float; coerce to
  Double before passing to round1(Double).
- BoxMeasureScreen: smart cast on delegated state.measuredL replaced with local
  val to satisfy Kotlin compiler.
- BoxMeasureScreen / QuoteInputSection: @OptIn(ExperimentalMaterial3Api::class)
  added for ExposedDropdownMenuBox / menuAnchor.
- HomeScreen (driver): Purple design token was missing; added.
- customer-app HomeScreen: COLORS.CYAN_NEON → COLORS.CYAN (token doesn't exist).

+++ apps/driver-app-android/feature/box-measure/src/main/kotlin/io/logisticos/driver/feature/boxmeasure/ui/ArCoreBoxMeasureView.kt
@@ -182,7 +182,7 @@ private class ArRenderer(
             val (p0, p1, p2, p3) = worldPts
             val l  = dist(p0, p1) * 100
             val w  = dist(p1, p2) * 100
-            val hh = abs(p3[1] - p2[1]) * 100
+            val hh = abs((p3[1] - p2[1]).toDouble()) * 100
             val confidence = when (frame.camera.trackingState) {
                 TrackingState.TRACKING -> 0.92
                 else -> 0.6
